+++ app/assets/javascripts/kolor-picker.js
@@ -6,16 +6,16 @@
   var colorPicker  = $();
 
   $(document).on('ready page:load', function () {
-    $('[data-kolor-picker]').each(function () {
+    $('[data-widget-kolor-picker]').each(function () {
       var element = $(this);
-      var options = element.data('kolor-picker') || {};
+      var options = element.data('widget-kolor-picker') || {};
       options.onSelect = function (color) {
         element.val(color);
       };
-      var colorWrapper = $('<div class="kolor-picker-color-wrapper" />');
-      var color = $('<div class="kolor-picker-color"/>').css('background-color', element.val() || 'white').appendTo(colorWrapper);
-      colorWrapper.insertAfter(element);
-      color.kolorPicker(options);
+      element.addClass('kolor-picker-input')
+        .attr('autocomplete', 'off')
+        .kolorPicker(options)
+        .wrap('<span class="kolor-picker-wrapper" />');
     });
   });