+++ lib/kolor-picker/builder.rb
@@ -4,7 +4,7 @@ module ActionView
   module Helpers
     class FormBuilder
       def colorpicker(method, options = {}, html_options = {})
-        @template.colorpicker(@object_name, method, options, html_options)
+        text_field(method, @template.colorpicker_options(options, html_options))
       end
     end
   end