+++ spec/helper_spec.rb
@@ -9,31 +9,31 @@ describe KolorPicker::Helper do
       context "with options" do
         context "with html_options" do
           subject { colorpicker_tag(:color, 'black', {:doRender => false}, {:class => :color}) }
-          it { should eq "<input type=\"hidden\" name=\"color\" id=\"color\" value=\"black\" class=\"color\" data-kolor-picker=\"{&quot;doRender&quot;:false}\" />" }
+          it { should eq "<input type=\"text\" name=\"color\" id=\"color\" value=\"black\" class=\"color\" data-widget-kolor-picker=\"{&quot;doRender&quot;:false}\" />" }
         end
 
         context "without html_options" do
           subject { colorpicker_tag(:color, 'black', :doRender => false) }
-          it { should eq "<input type=\"hidden\" name=\"color\" id=\"color\" value=\"black\" data-kolor-picker=\"{&quot;doRender&quot;:false}\" />" }
+          it { should eq "<input type=\"text\" name=\"color\" id=\"color\" value=\"black\" data-widget-kolor-picker=\"{&quot;doRender&quot;:false}\" />" }
         end
       end
 
       context "without options" do
         context "with html_options" do
           subject { colorpicker_tag(:color, 'black', nil, :class => :color) }
-          it { should eq "<input type=\"hidden\" name=\"color\" id=\"color\" value=\"black\" class=\"color\" data-kolor-picker=\"null\" />" }
+          it { should eq "<input type=\"text\" name=\"color\" id=\"color\" value=\"black\" class=\"color\" data-widget-kolor-picker=\"null\" />" }
         end
 
         context "without html_options" do
           subject { colorpicker_tag(:color, 'black') }
-          it { should eq "<input type=\"hidden\" name=\"color\" id=\"color\" value=\"black\" data-kolor-picker=\"{}\" />" }
+          it { should eq "<input type=\"text\" name=\"color\" id=\"color\" value=\"black\" data-widget-kolor-picker=\"{}\" />" }
         end
       end
     end
 
     context "with value" do
       subject { colorpicker_tag(:color) }
-      it { should eq "<input type=\"hidden\" name=\"color\" id=\"color\" data-kolor-picker=\"{}\" />" }
+      it { should eq "<input type=\"text\" name=\"color\" id=\"color\" data-widget-kolor-picker=\"{}\" />" }
     end
   end
 end