+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# v0.1.1 / 2016-05-10
+
+* Upgraded kolor-picker to v0.2.1
+* Fixed missing value.
+* Restructured helper and widget.
+
 # v0.1.0 / 2016-04-19
 
 Create project